add file with utils, extend hof-name-prefix with case when function has no brace
