+++ src/lib/rules/hof-name-prefix.ts
@@ -9,10 +9,7 @@
 
 import { TSESTree, TSESLint } from '@typescript-eslint/experimental-utils';
 
-type Function =
-  | TSESTree.FunctionDeclaration
-  | TSESTree.FunctionExpression
-  | TSESTree.ArrowFunctionExpression;
+import { isFunctionHigherOrder, getLastAncestor } from '../utils/ast-utils';
 
 type AllowedAncestors =
   | TSESTree.VariableDeclarator
@@ -30,10 +27,8 @@ const rule: TSESLint.RuleModule<string, string[]> = {
       url:
         'https://github.com/timon-and-pumbaa/eslint-plugin-fsd/blob/master/docs/rules/hof-name-prefix.md',
     },
-    fixable: undefined, // or "code" or "whitespace"
-    schema: [
-      // fill in your schema
-    ],
+    fixable: undefined,
+    schema: [],
     messages: {
       requireMake: 'Higher order functions name should start with make',
     },
@@ -56,37 +51,6 @@ const rule: TSESLint.RuleModule<string, string[]> = {
         node: identifier,
         messageId: 'requireMake',
       });
-    };
-
-    const isFunctionHigherOrder = (node: Function): boolean => {
-      if (!node.body) return false;
-
-      const functionBody = node.body as TSESTree.BlockStatement;
-      const { body: functionContent } = functionBody;
-
-      const functionTypes = ['ArrowFunctionExpression', 'FunctionExpression'];
-
-      const statementThatReturnsFunction = functionContent.find(
-        (contentNode: TSESTree.Node) => {
-          const nodeReturnsFunction =
-            contentNode.type === 'ReturnStatement' &&
-            contentNode.argument &&
-            functionTypes.includes(contentNode.argument.type);
-
-          if (nodeReturnsFunction) return true;
-
-          return false;
-        },
-      );
-
-      return !!statementThatReturnsFunction;
-    };
-
-    const getLastAncestor = (): TSESTree.Node => {
-      const ancestors = context.getAncestors();
-      const lastAncestor = ancestors[ancestors.length - 1];
-
-      return lastAncestor as TSESTree.Node;
     };
 
     const checkIdentifier = (identifier: TSESTree.Identifier): void => {
@@ -136,7 +100,7 @@ const rule: TSESLint.RuleModule<string, string[]> = {
         'Property',
       ];
 
-      const lastAncestor = getLastAncestor() as AllowedAncestors;
+      const lastAncestor = getLastAncestor(context) as AllowedAncestors;
       const functionIsHigherOrder = isFunctionHigherOrder(node);
 
       if (
